requirements.txt
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.0.0
tqdm>=4.60.0
tensorboard>=2.8.0
numpy>=1.21.0
